update: 员工管理完成 fix bug: mapper接口添加@Repository

+++ db/src/main/java/com/pyteam/db/mbg/mapper/Ab01Mapper.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface Ab01Mapper {
-    int countByExample(Ab01Example example);
+    long countByExample(Ab01Example example);
 
     int deleteByExample(Ab01Example example);
